Move `SequenceNumberCatchupBuffer to separate file
